Remove Delta from CLI
